feat(web): add post counts to blog category tabs
